refactor: 함수들 모듈 분리
검색 키워드 유효성 판단 함수 다른 파일로 분리, skeleton 템플릿 다른 파일로 분리, videoLIst 유무 판단 함수 클래스 내에 메서드로 이동 시킴
