+++ src/js/common/template.js
@@ -1,0 +1,9 @@
+const SKELETON_TEMPLATE = `
+  <div class="skeleton hidden">
+    <div class="image"></div>
+    <p class="line"></p>
+    <p class="line"></p>
+  </div>
+`;
+
+export { SKELETON_TEMPLATE };
